# openapi.snapshot.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FastAPI",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/campaigns/{campaign_id}/inventory/buy": {
      "post": {
        "tags": [
          "inventory"
        ],
        "summary": "Buy Item",
        "operationId": "buy_item_api_campaigns__campaign_id__inventory_buy_post",
        "parameters": [
          {
            "name": "campaign_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Campaign Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InventoryBuy"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InventoryRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/campaigns/{campaign_id}/members/me": {
      "get": {
        "tags": [
          "members"
        ],
        "summary": "Get Member Me",
        "operationId": "get_member_me_api_campaigns__campaign_id__members_me_get",
        "parameters": [
          {
            "name": "campaign_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Campaign Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MemberRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/campaigns/{campaign_id}/members/me/role": {
      "put": {
        "tags": [
          "members"
        ],
        "summary": "Update Member Role",
        "operationId": "update_member_role_api_campaigns__campaign_id__members_me_role_put",
        "parameters": [
          {
            "name": "campaign_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Campaign Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MemberRoleUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MemberRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "InventoryBuy": {
        "properties": {
          "itemId": {
            "type": "string",
            "title": "Itemid"
          },
          "quantity": {
            "type": "integer",
            "title": "Quantity",
            "default": 1
          }
        },
        "type": "object",
        "required": [
          "itemId"
        ],
        "title": "InventoryBuy"
      },
      "InventoryRead": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "campaignId": {
            "type": "string",
            "title": "Campaignid"
          },
          "memberId": {
            "type": "string",
            "title": "Memberid"
          },
          "itemId": {
            "type": "string",
            "title": "Itemid"
          },
          "quantity": {
            "type": "integer",
            "title": "Quantity"
          },
          "isEquipped": {
            "type": "boolean",
            "title": "Isequipped"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "title": "Createdat"
          },
          "updatedAt": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updatedat"
          }
        },
        "type": "object",
        "required": [
          "id",
          "campaignId",
          "memberId",
          "itemId",
          "quantity",
          "isEquipped",
          "notes",
          "createdAt",
          "updatedAt"
        ],
        "title": "InventoryRead"
      },
...
      "MemberRead": {
        "properties": {
          "campaignId": {
            "type": "string",
            "title": "Campaignid"
          },
          "displayName": {
            "type": "string",
            "title": "Displayname"
          },
          "roleMode": {
            "$ref": "#/components/schemas/RoleMode"
          }
        },
        "type": "object",
        "required": [
          "campaignId",
          "displayName",
          "roleMode"
        ],
        "title": "MemberRead"
      },
      "MemberRoleUpdate": {
        "properties": {
          "roleMode": {
            "$ref": "#/components/schemas/RoleMode"
          }
        },
        "type": "object",
        "required": [
          "roleMode"
        ],
        "title": "MemberRoleUpdate"
      },
...
      "RoleMode": {
        "type": "string",
        "enum": [
          "GM",
          "PLAYER"
        ],
        "title": "RoleMode"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}